backend/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
supabase
pydantic
requests
